+++ requirements.txt
@@ -6,3 +6,4 @@ pandas
 tushare
 git+git://github.com/zengbin93/cobra.git
 matplotlib
+zb>=0.1.14
